observability/backend-full-otel-lgtm/backends/ai-service/pyproject.toml
[tool.poetry]
name = "ai-service"
version = "0.1.0"
description = ""
authors = ["seungyeop-lee <xbeast7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
opentelemetry-instrumentation-fastapi = "^0.47b0"
opentelemetry-distro = "^0.47b0"
opentelemetry-api = "^1.26.0"
opentelemetry-sdk = "^1.26.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
